+++ index.js
@@ -1,29 +1,30 @@
 (function() {
-  var Storage, escape, fs, mkdirp, path, rimraf;
+  var Storage, encode, fs, mkdirp, path, rimraf;
 
   mkdirp = require('mkdirp');
 
-  rimraf = require('ramrif');
+  rimraf = require('rimraf');
 
   path = require('path');
 
   fs = require('fs');
 
-  escape = function(str) {
+  encode = function(str) {
     return encodeURIComponent(str).replace(/[!'()]/g, escape).replace(/\*/g, "%2A");
   };
 
   Storage = (function() {
     function Storage(dirname) {
       this.dirname = dirname;
-      mkdirp.sync(this.dirname);
     }
 
     Storage.prototype.getItem = function(key) {
       var file;
-      file = path.join(this.dirname, escape(key));
+      file = path.join(this.dirname, encode(key));
       if (fs.existsSync(file)) {
-        return fs.readFileSync(file);
+        return fs.readFileSync(file, {
+          encoding: 'UTF-8'
+        });
       } else {
         return null;
       }
@@ -31,18 +32,21 @@
 
     Storage.prototype.setItem = function(key, value) {
       var file;
-      file = path.join(this.dirname, escape(key));
+      mkdirp.sync(this.dirname);
+      file = path.join(this.dirname, encode(key));
       return fs.writeFileSync(file, value);
     };
 
     Storage.prototype.removeItem = function(key) {
       var file;
-      file = path.join(this.dirname, escape(key));
-      return fs.unlinkSync(file);
+      file = path.join(this.dirname, encode(key));
+      if (fs.existsSync(file)) {
+        return fs.unlinkSync(file);
+      }
     };
 
     Storage.prototype.clear = function() {
-      return ramrif.sync(this.dirname);
+      return rimraf.sync(this.dirname);
     };
 
     return Storage;
